Direct download buttons.
Rather than linking to download.php, link straight to the downloaded file and use JS to send a POST request to dl_click.php to track the download.
This is the same method that Texture Haven and 3D Model Haven uses. The advantage here is one less full page request, and no need to disable caching on download.php.